app/requirements.txt
streamlit
pandas
psycopg2-binary
matplotlib
wordcloud
python-dotenv
